refactor: removed test_templates
